(refactor) Move files into components and delete dashboard
