--- Dashboard.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
-
-const Dashboard = () => {
-  return (
-    <Text>Dashboard</Text>
-  );
-}
-
-export default Dashboard;
